WIP separate graphql Listing page from regular redux version, based on state
